chat_client/requirements.txt
chainlit>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
jinja2>=3.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
itsdangerous>=2.1.0
